src/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.6.4
numpy==1.26.4
pillow==10.2.0
torch==2.3.0
transformers==4.40.1
opencv-python==4.9.0.80
pymongo==4.6.2
redis==5.0.3
boto3==1.34.70
pandas
pyyaml
openai
huggingface-hub
timm
simple-parsing

